Optimization on debug grid drawing
TL&DR:
When you update a color to the material of the single square mesh, it redraws everytime. This is obviusly resource consuming because of the elevated number of SetPass calls.
(Reach to 250x250 square x time every frame)

The solution was to draw procedurally all the meshes and assign the correct premade material.

Previously they were 127849 etc. SetPass calls vs Now 4 pass calls. ( This is an huge performance optimization )
It reduce all the lag caused by istancing all these squares.

+ It resolve the problem where the grid is always under the generated mesh from marching.
+ Delegate the instancing of the grid to the gpu and remove useful handling from monobehaviour.